simplify api-server go files to one file, add go.mod
