+++ trunk/research/api-server/go.mod
@@ -1,0 +1,3 @@
+module api-server
+
+go 1.18
